+++ src/devices/advantium.ts
@@ -19,7 +19,7 @@ export class SmartHQAdvantium extends deviceBase {
     this.debugLog(`Advantium Features: ${JSON.stringify(accessory.context.device.features)}`)
 
     // Advantium Light
-    const light = this.accessory.getService('Advantium Light') ?? this.accessory.addService(this.platform.Service.Lightbulb, 'Advantium Light', 'AdvantiumLight')
+    const light = this.accessory!.getService('Advantium Light') ?? this.accessory!.addService(this.platform.Service.Lightbulb, 'Advantium Light', 'AdvantiumLight')
     light.setCharacteristic(this.platform.Characteristic.Name, 'Advantium Light')
     light
       .getCharacteristic(this.platform.Characteristic.On)
@@ -42,7 +42,7 @@ export class SmartHQAdvantium extends deviceBase {
       })
 
     // Advantium Temperature Sensor
-    const tempSensor = this.accessory.getService('Advantium Temperature') ?? this.accessory.addService(this.platform.Service.TemperatureSensor, 'Advantium Temperature', 'AdvantiumTemp')
+    const tempSensor = this.accessory!.getService('Advantium Temperature') ?? this.accessory!.addService(this.platform.Service.TemperatureSensor, 'Advantium Temperature', 'AdvantiumTemp')
     tempSensor.setCharacteristic(this.platform.Characteristic.Name, 'Advantium Temperature')
     tempSensor
       .getCharacteristic(this.platform.Characteristic.CurrentTemperature)
@@ -57,7 +57,7 @@ export class SmartHQAdvantium extends deviceBase {
       })
 
     // Advantium Running State (Switch)
-    const runningSwitch = this.accessory.getService('Advantium Running') ?? this.accessory.addService(this.platform.Service.Switch, 'Advantium Running', 'AdvantiumRunning')
+    const runningSwitch = this.accessory!.getService('Advantium Running') ?? this.accessory!.addService(this.platform.Service.Switch, 'Advantium Running', 'AdvantiumRunning')
     runningSwitch.setCharacteristic(this.platform.Characteristic.Name, 'Advantium Running')
     runningSwitch
       .getCharacteristic(this.platform.Characteristic.On)